third_party/blink/web_tests/webaudio/AudioContext/autoplay-explicit-suspension.html

<!DOCTYPE html>
<html>
  <head>
    <title>
      Test if starting source node does not resume an explicitly suspended
      context
    </title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      const emulateUserGesture = () => {
        return new Promise((resolve, reject) => {
          chrome.gpuBenchmarking.pointerActionSequence([{
            source: 'mouse',
            actions: [
              { name: 'pointerDown', x: 1, y: 1 },
              { name: 'pointerUp' },
            ]
          }], resolve);
        });
      };

      promise_test(async t => {
        const autoplayFlag = internals.settings.autoplayIgnoresWebAudioEnabled;

        internals.settings.setAutoplayPolicy(
            'document-user-activation-required');
        internals.settings.autoplayIgnoresWebAudioEnabled = false;

        t.add_cleanup(() => {
          internals.settings.autoplayIgnoresWebAudioEnabled = autoplayFlag;
          internals.settings.setAutoplayPolicy('no-user-gesture-required');
        });

        // Suspend a context explicitly and then start a source node after
        // user gesture. The context state should stay the same.
        const context = new AudioContext();
        const source = new OscillatorNode(context);
        source.connect(context.destination);

        await context.suspend();

        assert_equals(
            context.state,
            'suspended',
            'The context state after explicit suspension');

        // Starting a source node should not unlock an explicitly suspended
        // context, even if the event is originated from user gesture.
        await emulateUserGesture();
        source.start();

        assert_equals(
            context.state,
            'suspended',
            'The context state after user gesture followed by source start');
      }, 'Test if starting source node does not resume an explicitly ' +
          'suspended context');
    </script>
  </body>
</html>
